Authorize PartController

+++ app/Http/Controllers/PartController.php
@@ -4,12 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Part;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class PartController extends Controller
 {
     public function index()
     {
+        Gate::authorize('parts');
+
         $parts = Part::query();
 
         if ($keyword = request('search')) {
@@ -29,11 +32,15 @@ class PartController extends Controller
 
     public function create()
     {
+        Gate::authorize('parts');
+
         return view('parts.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('parts');
+
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'unit' => 'required|string|max:255',
@@ -55,15 +62,19 @@ class PartController extends Controller
 
     public function edit(Part $part)
     {
+        Gate::authorize('parts');
+
         return view('parts.edit', compact('part'));
     }
 
     public function update(Request $request, Part $part)
     {
+        Gate::authorize('parts');
+
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'unit' => 'required|string|max:255',
-            'code' => ['required','numeric',Rule::unique('parts')->ignore($part->id)],
+            'code' => ['required', 'numeric', Rule::unique('parts')->ignore($part->id)],
             'price' => 'nullable'
         ]);
 
@@ -76,6 +87,8 @@ class PartController extends Controller
 
     public function destroy(Part $part)
     {
+        Gate::authorize('parts');
+
         $part->delete();
 
         alert()->success('حذف موفق', 'حذف قطعه با موفقیت انجام شد');
